fix: critical issues & code quality improvements for v1
Remove debug logs, fix token storage, and passkey state issues; create shared blob utility.
